Quality Exchange
What does a quality crypto exchange look like, in the modern world, as we head towards 2020? An exchange is a website that has been designed, so that an end user can buy, sell or exchange crypto for fiat, such as the USD or Euro. It is also a primary stage for traders to have a professional experience with advanced trading tools at their fingers. Over the recent years, many exchanges require the user to verify their identification in order to open an account, which gives them admittance to the exchange.

There are at least two types of known exchanges. The first being a trading platform, while the second, being a direct trading platform. A trading platform connects buyers and sellers and charges a standard fee for each transaction that takes place on that exchange. The ProperSix exchange would be considered a trading platform compared to a direct trading platform. Whereas, direct trading platforms allow the sellers to set an exchange rate that they feel is sufficient to the transaction.
Then token/coin pairing arises. Many exchanges will offer you the popular BTC or ETH options. Although, there are so many more crypto selections to choose from. The more variety and diversification offered by an exchange sets the user up for a lively experience. ProperSix doesn’t have a complete list of coins/tokens being offered currently online just yet. Once the exchange is live, lookout for this information to be released by the ProperSix team!
Liquidity is based off of the buying and selling ease of the given exchange. Having high liquidity means the exchange has a fairly large group of sellers and buyers. This can lead to faster transactions and better pricing within the exchange. I have a feeling the ProperSix exchange will boom with new business once officially live!

Customer Support and the quality of such that an exchange offers is a biggie! A great support team will be extremely responsive to the customer if they need help related to identification verification processes, withdrawals, deposits or trading orders. These are some examples that a customer service member may be required to perform. Slow responses can cost the user a lot of fiat and even more in loss time. ProperSix, thus far seems to be very responsive to their telegram channel and website traffic, which is positive.
Trading fees become a sticky mess at times for exchanges. Higher fees will eat your profit margins away without doubt. It is important to evaluate said fees prior to initiating a deposit or withdraw. ProperSix exchange seems to be fair in comparison to others. Check them out!
Lastly, ease of use of an exchange should create very little confusion for the user. It should create a clear and easy familiarity for the customer, while navigating the exchange platform. The ProperSix team has developed their exchange to be such an advanced trading platform, any trader can’t go wrong!

https://www.propersix.com